requirements.txt
requests>=2.31.0
PyYAML>=6.0.1
# Required when using auth.type: playwright_login
playwright>=1.47.0
